Refactored core admin into a module
