+++ core/admin/user.py
@@ -1,0 +1,56 @@
+from django.contrib import admin
+from django.contrib.auth import admin as auth_admin
+from django.contrib.auth.forms import UserChangeForm
+
+from core.admin.forms.user import UserLimitedChangeForm, UserCreationForm
+
+
+class UserAdmin(auth_admin.UserAdmin):
+    fieldsets = (
+        (None, {'fields': ('email', 'password')}),
+        ('Personal info', {'fields': ('first_name', 'last_name')}),
+        ('Permissions', {'fields': ('is_active', 'is_staff', 'is_superuser',
+                                    'groups', 'user_permissions')}),
+        ('Important dates', {'fields': ('last_login', 'date_joined')}),
+    )
+    limited_fieldsets = (
+        (None, {'fields': ('email',)}),
+        ('Personal info', {'fields': ('first_name', 'last_name')}),
+        ('Important dates', {'fields': ('last_login', 'date_joined')}),
+    )
+    add_fieldsets = (
+        (None, {
+            'classes': ('wide',),
+            'fields': ('email', 'password1', 'password2')
+        }),
+    )
+    form = UserChangeForm
+    limited_form = UserLimitedChangeForm
+    add_form = UserCreationForm
+    change_password_form = auth_admin.AdminPasswordChangeForm
+    list_display = ('email', 'first_name', 'last_name', 'is_superuser', 'date_joined')
+    list_filter = ('event', 'is_staff', 'is_superuser', 'is_active', 'groups', 'date_joined')
+    search_fields = ('first_name', 'last_name', 'email')
+    ordering = ('email',)
+    readonly_fields = ('last_login', 'date_joined',)
+
+    def get_queryset(self, request):
+        qs = super(UserAdmin, self).get_queryset(request)
+        if request.user.is_superuser:
+            return qs
+        return qs.filter(pk=request.user.pk)
+
+    def get_form(self, request, obj=None, **kwargs):
+        defaults = {}
+        if obj and not request.user.is_superuser:
+            defaults.update({
+                'form': self.limited_form,
+                'fields': admin.util.flatten_fieldsets(self.limited_fieldsets),
+            })
+        defaults.update(kwargs)
+        return super(UserAdmin, self).get_form(request, obj, **defaults)
+
+    def get_fieldsets(self, request, obj=None):
+        if obj and not request.user.is_superuser:
+            return self.limited_fieldsets
+        return super(UserAdmin, self).get_fieldsets(request, obj)
